requirements.txt
aiofiles>=25.1.0
cloudinary>=1.41.0
email-validator>=2.3.0
fastapi>=0.123.9
itsdangerous>=2.2.0
jinja2>=3.1.6
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pillow>=12.0.0
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1
python-jose[cryptography]>=3.5.0
python-multipart>=0.0.20
sqlalchemy>=2.0.44
uvicorn[standard]>=0.38.0
werkzeug>=3.1.4
